Deprecate older pages for my view report and view team report pages
