+++ src/app/fyle/fyle-routing.module.ts
@@ -32,11 +32,6 @@ const routes: Routes = [
   {
     path: 'my_view_report',
     loadChildren: () => import('./my-view-report/my-view-report.module').then((m) => m.MyViewReportPageModule),
-    canActivate: [BetaPageFeatureFlagGuard],
-  },
-  {
-    path: 'my_view_report_beta',
-    loadChildren: () => import('./my-view-report/my-view-report-v2.module').then((m) => m.MyViewReportV2PageModule),
   },
   {
     path: 'help',
@@ -62,12 +57,6 @@ const routes: Routes = [
   {
     path: 'view_team_report',
     loadChildren: () => import('./view-team-report/view-team-report.module').then((m) => m.ViewTeamReportPageModule),
-    canActivate: [BetaPageFeatureFlagGuard],
-  },
-  {
-    path: 'view_team_report_beta',
-    loadChildren: () =>
-      import('./view-team-report/view-team-report-v2.module').then((m) => m.ViewTeamReportPageV2Module),
   },
   {
     path: 'my_view_advance',
