--- src/app/fyle/my-view-report/add-expenses-to-report-v2/add-expenses-to-report-v2.component.html
@@ -1,85 +1,0 @@
-<ion-header mode="md">
-  <mat-icon class="fy-modal-top-bar" svgIcon="fy-rectangle"></mat-icon>
-  <ion-toolbar mode="md" class="fy-modal-toolbar">
-    <ion-title>
-      <div *ngIf="selectedElements?.length === 0" class="report-list--title text-center">Add Expenses</div>
-      <div *ngIf="selectedElements?.length > 0" class="add-expenses-to-report--title-container text-center">
-        <div class="add-expenses-to-report--title">
-          {{ selectedElements?.length }} {{ selectedElements?.length > 1 ? 'Expenses' : 'Expense' }} -
-          {{ selectedTotalAmount || 0 | humanizeCurrency : homeCurrency }}
-        </div>
-      </div>
-    </ion-title>
-    <ion-buttons slot="start">
-      <ion-button (click)="close()">
-        <mat-icon class="fy-icon-close" svgIcon="fy-close"></mat-icon>
-      </ion-button>
-    </ion-buttons>
-    <ion-buttons slot="end">
-      <ion-button (click)="addNewExpense()">
-        <ion-icon class="report-list--add-icon" src="../../../../../assets/svg/plus-square.svg"></ion-icon>
-      </ion-button>
-    </ion-buttons>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content>
-  <div class="fy-modal">
-    <div *ngIf="unreportedExpenses?.length > 0" class="add-expenses-to-report--select-all-checkbox">
-      <mat-checkbox
-        class="custom-mat-checkbox create-new-report--checkbox"
-        [(ngModel)]="isSelectedAll"
-        (ngModelChange)="toggleSelectAll($event)"
-        name="select all"
-      >
-        Select all</mat-checkbox
-      >
-    </div>
-
-    <div *ngFor="let expense of unreportedExpenses as list; let i = index">
-      <app-expense-card-v2
-        [expense]="expense"
-        [expenseIndex]="i"
-        [previousExpenseTxnDate]="list[i - 1]?.spent_at"
-        [previousExpenseCreatedAt]="list[i - 1]?.created_at"
-        [isSelectionModeEnabled]="true"
-        [selectedElements]="selectedElements"
-        (cardClickedForSelection)="toggleExpense(expense)"
-        [isFromViewReports]="true"
-      >
-      </app-expense-card-v2>
-    </div>
-  </div>
-  <ng-container *ngIf="unreportedExpenses?.length === 0">
-    <div class="add-expenses-to-report--zero-state">
-      <div class="text-center">
-        <img src="../../../assets/images/zero-states/expenses.png" alt="No expense in this report" />
-        <div class="add-expenses-to-report--zero-state--header text-center">
-          Looks like there are no complete expenses!
-        </div>
-
-        <div class="add-expenses-to-report--zero-state--sub-header">
-          Click on the
-          <mat-icon class="add-expenses-to-report--zero-state--sub-header--icon" svgIcon="plus-square"></mat-icon>
-          to add a new expense to this report
-        </div>
-      </div>
-    </div>
-  </ng-container>
-</ion-content>
-
-<ion-footer>
-  <ion-toolbar mode="md" class="add-expenses-to-report--footer">
-    <div>
-      <div class="fy-footer-cta-container">
-        <ion-button
-          class="fy-footer-cta fy-footer-cta--primary"
-          [disabled]="!(selectedTotalExpenses > 0)"
-          (click)="addExpensestoReport()"
-        >
-          Add to Report
-        </ion-button>
-      </div>
-    </div>
-  </ion-toolbar>
-</ion-footer>
