--- src/app/fyle/my-view-report/add-expenses-to-report-v2/add-expenses-to-report-v2.component.scss
@@ -1,68 +1,0 @@
-@import '.././../../../theme/colors.scss';
-
-.add-expenses-to-report {
-  &--toolbar {
-    background-color: $pure-white;
-    border-bottom: 1px solid $grey-lighter;
-  }
-
-  &--title-container {
-    color: $black;
-    font-weight: 500;
-    margin-left: -10%;
-  }
-
-  &--title {
-    font-size: 20px;
-    line-height: 26px;
-  }
-
-  &--close {
-    margin-left: 16px;
-  }
-
-  &--select-all-checkbox {
-    font-size: 14px;
-    padding: 16px 16px;
-    background-color: $pure-white;
-  }
-
-  &--zero-state {
-    font-size: 14px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    color: $blue-black;
-    line-height: 1.5;
-    font-weight: 500;
-    margin: 20% 20px 10px;
-
-    &--header {
-      margin-top: 12px;
-      margin-bottom: 10px;
-      font-size: 15px;
-      color: $black;
-      font-weight: 500;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-    }
-    &--sub-header {
-      font-size: 14px;
-      text-align: center;
-      color: $dark-grey;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: -1px;
-      &--icon {
-        height: 14px;
-        color: $dark-grey;
-      }
-    }
-  }
-
-  &--footer {
-    box-shadow: 0px -2px 40px rgba(215, 215, 215, 0.4);
-  }
-}
