+++ src/app/fyle/my-view-report/add-expenses-to-report/add-expenses-to-report.component.html
@@ -25,7 +25,7 @@
 
 <ion-content>
   <div class="fy-modal">
-    <div *ngIf="unReportedEtxns?.length > 0" class="add-expenses-to-report--select-all-checkbox">
+    <div *ngIf="unreportedExpenses?.length > 0" class="add-expenses-to-report--select-all-checkbox">
       <mat-checkbox
         class="custom-mat-checkbox create-new-report--checkbox"
         [(ngModel)]="isSelectedAll"
@@ -36,21 +36,21 @@
       >
     </div>
 
-    <div *ngFor="let etxn of unReportedEtxns as list; let i = index">
-      <app-expense-card
-        [expense]="etxn"
-        [etxnIndex]="i"
-        [previousExpenseTxnDate]="list[i - 1]?.tx_txn_dt"
-        [previousExpenseCreatedAt]="list[i - 1]?.tx_created_at"
+    <div *ngFor="let expense of unreportedExpenses as list; let i = index">
+      <app-expense-card-v2
+        [expense]="expense"
+        [expenseIndex]="i"
+        [previousExpenseTxnDate]="list[i - 1]?.spent_at"
+        [previousExpenseCreatedAt]="list[i - 1]?.created_at"
         [isSelectionModeEnabled]="true"
         [selectedElements]="selectedElements"
-        (cardClickedForSelection)="toggleTransaction(etxn)"
+        (cardClickedForSelection)="toggleExpense(expense)"
         [isFromViewReports]="true"
       >
-      </app-expense-card>
+      </app-expense-card-v2>
     </div>
   </div>
-  <ng-container *ngIf="unReportedEtxns?.length === 0">
+  <ng-container *ngIf="unreportedExpenses?.length === 0">
     <div class="add-expenses-to-report--zero-state">
       <div class="text-center">
         <img src="../../../assets/images/zero-states/expenses.png" alt="No expense in this report" />
@@ -74,7 +74,7 @@
       <div class="fy-footer-cta-container">
         <ion-button
           class="fy-footer-cta fy-footer-cta--primary"
-          [disabled]="!(selectedTotalTxns > 0)"
+          [disabled]="!(selectedTotalExpenses > 0)"
           (click)="addExpensestoReport()"
         >
           Add to Report
