--- src/app/fyle/my-view-report/edit-report-name-popover-v2/edit-report-name-popover.component.scss
@@ -1,49 +1,0 @@
-@import '../../../../theme/colors.scss';
-
-.edit-report-name {
-  &--toolbar {
-    &__title {
-      color: $black;
-      padding-left: 20px;
-    }
-
-    &__btn-save {
-      font-size: 16px;
-      line-height: 1.3;
-      color: $brand-primary;
-      text-transform: none;
-    }
-  }
-
-  &--input-container {
-    padding: 16px 16px 36px;
-
-    &__label {
-      margin: 0 8px 0 0;
-      font-size: 12px;
-      color: $black-light;
-      line-height: 1.3;
-      font-weight: 400;
-    }
-
-    &__input {
-      border: 0;
-      font-size: 14px;
-      font-weight: 400;
-      line-height: 1.3;
-      color: $blue-black;
-      width: 100%;
-      padding: 6px 0;
-      border-bottom: 1px solid $grey-lighter;
-    }
-
-    &__mandatory {
-      color: $brand-primary;
-      font-size: 14px;
-    }
-
-    &__input:focus {
-      border-bottom: 1px solid $blue-black;
-    }
-  }
-}
