+++ src/app/fyle/my-view-report/my-view-report.page.html
@@ -159,17 +159,17 @@
           <div *ngIf="!isExpensesLoading && erpt.rp_num_transactions > 0">
             <div class="view-reports--divider"></div>
             <div class="view-reports--expenses-card-block">
-              <div *ngFor="let etxn of etxns$|async as list; let i = index;">
-                <app-expense-card
-                  [expense]="etxn"
-                  [etxnIndex]="i"
-                  [previousExpenseTxnDate]="list[i-1]?.tx_txn_dt"
-                  [previousExpenseCreatedAt]="list[i-1]?.tx_created_at"
+              <div *ngFor="let expense of expenses$|async as list; let i = index;">
+                <app-expense-card-v2
+                  [expense]="expense"
+                  [expenseIndex]="i"
+                  [previousExpenseTxnDate]="list[i-1]?.spent_at"
+                  [previousExpenseCreatedAt]="list[i-1]?.created_at"
                   (goToTransaction)="goToTransaction($event)"
                   [isFromReports]="true"
                   [isFromViewReports]="true"
                 >
-                </app-expense-card>
+                </app-expense-card-v2>
               </div>
             </div>
             <ng-template #ExpensesLoader>
@@ -181,7 +181,7 @@
               <img src="../../../assets/images/zero-states/expenses.png" alt="No expense in this report" />
               <div class="view-reports--zero-state-header">Looks like this report is empty.</div>
               <button
-                *ngIf="unReportedEtxns && unReportedEtxns.length > 0"
+                *ngIf="unreportedExpenses && unreportedExpenses.length > 0"
                 class="view-reports--unreported-cta"
                 (click)="showAddExpensesToReportModal()"
               >
@@ -315,7 +315,7 @@
 >
   <ion-toolbar class="cta-footer fy-footer-cta-container" mode="md">
     <div class="view-reports--cta-container">
-      <div class="view-reports--add-expenses-container" *ngIf="unReportedEtxns">
+      <div class="view-reports--add-expenses-container" *ngIf="unreportedExpenses">
         <div class="view-reports--add-expenses-container__icon_container">
           <mat-icon
             (click)="showAddExpensesToReportModal()"
@@ -345,7 +345,7 @@
 
 <ion-footer
   class="view-reports--add-expense-footer"
-  *ngIf="segmentValue === Segment.EXPENSES && ((erpt$|async)?.rp_state === 'APPROVER_PENDING') && unReportedEtxns"
+  *ngIf="segmentValue === Segment.EXPENSES && ((erpt$|async)?.rp_state === 'APPROVER_PENDING') && unreportedExpenses"
 >
   <div class="view-reports--add-expense-footer__block">
     <div class="view-reports--add-expense-footer__cta" (click)="showAddExpensesToReportModal()">
