--- src/app/fyle/my-view-report/share-report-v2/share-report.component.html
@@ -1,36 +1,0 @@
-<ion-header mode="md">
-  <mat-icon class="fy-modal-top-bar" svgIcon="fy-rectangle"></mat-icon>
-  <ion-toolbar mode="md" class="text-center fy-modal-toolbar share-report--toolbar">
-    <ion-title>
-      <div class="share-report--toolbar__title">Share Report</div>
-    </ion-title>
-    <ion-buttons slot="start">
-      <ion-button (click)="cancel()">
-        <mat-icon class="fy-icon-close" svgIcon="fy-close"></mat-icon>
-      </ion-button>
-    </ion-buttons>
-    <ion-buttons slot="end">
-      <ion-button
-        [disabled]="!emailInput.value || emailInput.invalid"
-        class="share-report--toolbar__btn-share"
-        (click)="shareReport(emailInput)"
-      >
-        Share
-      </ion-button>
-    </ion-buttons>
-  </ion-toolbar>
-</ion-header>
-
-<div class="share-report--input-container">
-  <div class="share-report--input-container__label">Share report with</div>
-  <input
-    #simpleEmailInput
-    matInput
-    type="email"
-    placeholder="Email ID"
-    [(ngModel)]="email"
-    pattern="\S+@\S+\.\S{2,}"
-    #emailInput="ngModel"
-    class="share-report--input-container__input smartlook-show"
-  />
-</div>
