--- src/app/fyle/my-view-report/share-report-v2/share-report.component.scss
@@ -1,44 +1,0 @@
-@import '../../../../theme/colors.scss';
-
-.share-report {
-  &--toolbar {
-    &__title {
-      color: $black;
-      padding-left: 38px;
-    }
-
-    &__btn-share {
-      font-size: 16px;
-      line-height: 1.3;
-      color: $brand-primary;
-      text-transform: none;
-    }
-  }
-
-  &--input-container {
-    padding: 16px 16px 36px;
-
-    &__label {
-      margin: 0 8px 0 0;
-      font-size: 12px;
-      color: $black-light;
-      line-height: 1.3;
-      font-weight: 400;
-    }
-
-    &__input {
-      border: 0;
-      font-size: 14px;
-      font-weight: 400;
-      line-height: 1.3;
-      color: $blue-black;
-      width: 100%;
-      padding: 6px 0;
-      border-bottom: 1px solid $grey-lighter;
-    }
-
-    &__input:focus {
-      border-bottom: 1px solid $blue-black;
-    }
-  }
-}
