--- src/app/fyle/view-team-report/view-team-report-v2-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { ViewTeamReportPageV2 } from './view-team-report-v2.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: ViewTeamReportPageV2,
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class ViewTeamReportPageV2RoutingModule {}
