+++ src/app/fyle/view-team-report/view-team-report.module.ts
@@ -2,15 +2,15 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
-import { ViewTeamReportPageRoutingModule } from './view-team-report-routing.module';
-import { ViewTeamReportPage } from './view-team-report.page';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatRippleModule } from '@angular/material/core';
 import { SharedModule } from 'src/app/shared/shared.module';
-import { ShareReportComponent } from './share-report/share-report.component';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { ViewTeamReportPage } from './view-team-report.page';
+import { ViewTeamReportPageRoutingModule } from './view-team-report-routing.module';
+import { ShareReportComponent } from './share-report/share-report.component';
 
 @NgModule({
   imports: [
