--- src/app/shared/components/fy-view-report-info-v2/fy-view-report-info.component.html
@@ -1,119 +1,0 @@
-<ion-header>
-  <mat-icon *ngIf="platform.is('ios')" class="fy-modal-top-bar" svgIcon="fy-rectangle"></mat-icon>
-  <ion-toolbar mode="md" class="text-center fy-modal-toolbar view-info--toolbar">
-    <ion-buttons slot="start">
-      <ion-button (click)="closeModal()">
-        <mat-icon class="fy-icon-close" svgIcon="fy-close"></mat-icon>
-      </ion-button>
-    </ion-buttons>
-    <ion-title>
-      <div class="view-info--toolbar__title">View Info</div>
-    </ion-title>
-  </ion-toolbar>
-
-  <ion-toolbar class="view-info--segment-block-container">
-    <ion-segment
-      class="segment-block view-info--segment-block-container__segment"
-      (ionChange)="segmentChanged($event)"
-      value="report"
-    >
-      <ion-segment-button value="report" class="view-info--segment-block-container__btn text-capitalize">
-        Report
-      </ion-segment-button>
-      <ion-segment-button value="amount" class="view-info--segment-block-container__btn text-capitalize">
-        Amount
-      </ion-segment-button>
-      <ion-segment-button
-        *ngIf="view === ExpenseView.team"
-        value="employee"
-        class="view-info--segment-block-container__btn text-capitalize"
-      >
-        Employee
-      </ion-segment-button>
-    </ion-segment>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content *ngIf="isReportView && !isEmployeeView" (swipe)="onSwipeReport($event)">
-  <ion-grid class="view-info--body-container">
-    <ion-row *ngFor="let item of reportDetails | keyvalue : originalOrder" class="view-info--body-container__content">
-      <ion-col size="5">
-        <div class="view-info--body-container__title">
-          {{ item.key }}
-        </div>
-      </ion-col>
-      <ion-col size="7">
-        <div class="view-info--body-container__value">
-          {{ item.value }}
-        </div>
-      </ion-col>
-    </ion-row>
-  </ion-grid>
-</ion-content>
-
-<ion-content *ngIf="!isReportView && !isEmployeeView" (swipe)="onSwipeAmount($event)">
-  <p class="view-info--section-header">COMPONENT WISE SPLIT</p>
-  <ion-grid class="view-info--body-container">
-    <ion-row
-      *ngFor="let item of amountComponentWiseDetails | keyvalue : originalOrder"
-      class="view-info--body-container__content"
-    >
-      <ion-col size="5">
-        <div class="view-info--body-container__title">
-          {{ item.key }}
-        </div>
-        <div *ngIf="item && (item.key === 'Advance' || item.key === 'CCC')" class="view-info--body-container__subtitle">
-          Non-Reimbursable
-        </div>
-      </ion-col>
-      <ion-col size="7">
-        <div class="view-info--body-container__value">
-          {{ item.value | currency : reportCurrency : 'symbol-narrow' }}
-        </div>
-      </ion-col>
-    </ion-row>
-  </ion-grid>
-  <div *ngIf="isForeignCurrency">
-    <p class="view-info--section-header view-info--section-header__currency">CURRENCY WISE SPLIT</p>
-    <ion-grid class="view-info--body-container">
-      <ion-row *ngFor="let item of amountCurrencyWiseDetails" class="view-info--body-container__content">
-        <ion-col size="5">
-          <div class="view-info--body-container__title">
-            {{ item.name }}
-          </div>
-          <div class="view-info--body-container__subtitle">
-            {{ item.count === 1 ? '1 Expense' : item.count + ' Expenses' }}
-          </div>
-        </ion-col>
-        <ion-col size="7">
-          <div class="view-info--body-container__value">
-            <ng-container *ngIf="item.currency !== reportCurrency; else origAmountOnly">
-              {{ item.origAmount | currency : item.currency : 'code' }} =
-              {{ item.amount | currency : reportCurrency : 'code' }}
-            </ng-container>
-            <ng-template #origAmountOnly> {{ item.origAmount | currency : item.currency : 'code' }} </ng-template>
-          </div>
-        </ion-col>
-      </ion-row>
-    </ion-grid>
-  </div>
-</ion-content>
-
-<ion-content *ngIf="view === ExpenseView.team && isEmployeeView && !isReportView" (swipe)="onSwipeEmployee($event)">
-  <ion-grid class="view-info--body-container">
-    <ng-container *ngFor="let item of employeeDetails | keyvalue : originalOrder">
-      <ion-row *ngIf="item && item.value && item.value.length" class="view-info--body-container__content">
-        <ion-col size="5">
-          <div class="view-info--body-container__title">
-            {{ item.key }}
-          </div>
-        </ion-col>
-        <ion-col size="7">
-          <div class="view-info--body-container__value">
-            {{ item.value }}
-          </div>
-        </ion-col>
-      </ion-row>
-    </ng-container>
-  </ion-grid>
-</ion-content>
