--- src/app/shared/components/fy-view-report-info-v2/fy-view-report-info.component.scss
@@ -1,73 +1,0 @@
-@import '../../../../theme/colors.scss';
-
-.view-info {
-  &--toolbar {
-    text-align: center;
-    margin-bottom: 12px;
-
-    &__title {
-      color: $black;
-      padding-right: 44px;
-    }
-  }
-
-  &--segment-block-container {
-    --border-width: 0 !important;
-    --background: #{$pure-white};
-
-    &__segment.ios {
-      margin: 12px auto;
-    }
-
-    &__segment.md {
-      padding: 0 16px 8px;
-    }
-
-    &__btn {
-      color: $blue-black;
-      font-size: 14px;
-      font-weight: 500;
-      width: 100%;
-    }
-  }
-
-  &--section-header {
-    font-size: 14px;
-    font-weight: 500;
-    background: $white;
-    height: 40px;
-    padding: 12px 16px;
-    margin: 0;
-    color: $grey-light;
-
-    &__currency {
-      margin-top: -1px;
-    }
-  }
-
-  &--body-container {
-    padding: 0 16px;
-    font-size: 16px;
-    line-height: 1.3;
-    color: $black-light;
-
-    &__content {
-      min-height: 48px;
-      border-bottom: 1px solid $grey;
-      padding: 12px 0;
-    }
-
-    &__title {
-      font-weight: 500;
-    }
-
-    &__subtitle {
-      padding-top: 4px;
-      font-size: 12px;
-    }
-
-    &__value {
-      color: $black;
-    }
-  }
-}
