+++ src/app/shared/components/fy-view-report-info/fy-view-report-info.component.spec.ts
@@ -2,10 +2,8 @@ import { ComponentFixture, TestBed, fakeAsync, tick, waitForAsync } from '@angul
 
 import { FyViewReportInfoComponent } from './fy-view-report-info.component';
 import { reportParam } from 'src/app/core/mock-data/report.data';
-import { expenseList } from 'src/app/core/mock-data/expense.data';
 import { of } from 'rxjs';
 import { ExpenseView } from 'src/app/core/models/expense-view.enum';
-import { TransactionService } from 'src/app/core/services/transaction.service';
 import { OrgUserSettingsService } from 'src/app/core/services/org-user-settings.service';
 import { TrackingService } from 'src/app/core/services/tracking.service';
 import { AuthService } from 'src/app/core/services/auth.service';
@@ -18,11 +16,13 @@ import { currencySummaryData } from 'src/app/core/mock-data/currency-summary.dat
 import { apiEouRes } from 'src/app/core/mock-data/extended-org-user.data';
 import { costCentersData } from 'src/app/core/mock-data/cost-centers.data';
 import { cloneDeep } from 'lodash';
+import { expenseResponseData, expenseResponseData2 } from 'src/app/core/mock-data/platform/v1/expense.data';
+import { ExpensesService as SharedExpensesService } from 'src/app/core/services/platform/v1/shared/expenses.service';
 
 describe('FyViewReportInfoComponent', () => {
   let component: FyViewReportInfoComponent;
   let fixture: ComponentFixture<FyViewReportInfoComponent>;
-  let transactionService: jasmine.SpyObj<TransactionService>;
+  let sharedExpensesService: jasmine.SpyObj<SharedExpensesService>;
   let orgUserSettingsService: jasmine.SpyObj<OrgUserSettingsService>;
   let trackingService: jasmine.SpyObj<TrackingService>;
   let authService: jasmine.SpyObj<AuthService>;
@@ -31,13 +31,13 @@ describe('FyViewReportInfoComponent', () => {
   let datePipe: DatePipe;
 
   beforeEach(waitForAsync(() => {
-    const mockTransactionServiceSpy = jasmine.createSpyObj('TransactionService', [
-      'getPaymentModeWiseSummary',
-      'getCurrenyWiseSummary',
-    ]);
     const mockOrgUserSettingsServiceSpy = jasmine.createSpyObj('OrgUserSettingsService', [
       'get',
       'getAllowedCostCentersByOuId',
+    ]);
+    const mockSharedExpensesServiceSpy = jasmine.createSpyObj('SharedExpensesService', [
+      'getPaymentModeWiseSummary',
+      'getCurrenyWiseSummary',
     ]);
     const mockTrackingServiceSpy = jasmine.createSpyObj('TrackingService', ['viewReportInfo']);
     const mockAuthServiceSpy = jasmine.createSpyObj('AuthService', ['getUserDetails', 'getEou']);
@@ -48,8 +48,8 @@ describe('FyViewReportInfoComponent', () => {
       declarations: [FyViewReportInfoComponent],
       providers: [
         {
-          provide: TransactionService,
-          useValue: mockTransactionServiceSpy,
+          provide: SharedExpensesService,
+          useValue: mockSharedExpensesServiceSpy,
         },
         {
           provide: OrgUserSettingsService,
@@ -79,7 +79,7 @@ describe('FyViewReportInfoComponent', () => {
 
     fixture = TestBed.createComponent(FyViewReportInfoComponent);
     component = fixture.componentInstance;
-    transactionService = TestBed.inject(TransactionService) as jasmine.SpyObj<TransactionService>;
+    sharedExpensesService = TestBed.inject(SharedExpensesService) as jasmine.SpyObj<SharedExpensesService>;
     datePipe = TestBed.inject(DatePipe);
     orgUserSettingsService = TestBed.inject(OrgUserSettingsService) as jasmine.SpyObj<OrgUserSettingsService>;
     trackingService = TestBed.inject(TrackingService) as jasmine.SpyObj<TrackingService>;
@@ -120,10 +120,10 @@ describe('FyViewReportInfoComponent', () => {
     spyOn(component, 'createEmployeeDetails');
     spyOn(component, 'getCCCAdvanceSummary');
     orgSettingsService.get.and.returnValue(of(orgSettingsRes));
-    transactionService.getCurrenyWiseSummary.and.returnValue(currencySummaryData);
-    transactionService.getPaymentModeWiseSummary.and.returnValue(paymentModeSummaryMock);
+    sharedExpensesService.getCurrenyWiseSummary.and.returnValue(currencySummaryData);
+    sharedExpensesService.getPaymentModeWiseSummary.and.returnValue(paymentModeSummaryMock);
     component.erpt$ = of(reportParam);
-    component.etxns$ = of(expenseList);
+    component.expenses$ = of(expenseResponseData);
     fixture.detectChanges();
     component.ionViewWillEnter();
     expect(component.reportDetails).toEqual(erpt);
@@ -134,12 +134,12 @@ describe('FyViewReportInfoComponent', () => {
       Reimbursable: 4600,
     });
     expect(component.getCCCAdvanceSummary).toHaveBeenCalledOnceWith(paymentModeSummaryMock, orgSettingsRes);
-    expect(transactionService.getCurrenyWiseSummary).toHaveBeenCalledOnceWith(expenseList);
+    expect(sharedExpensesService.getCurrenyWiseSummary).toHaveBeenCalledOnceWith(expenseResponseData);
     expect(component.amountCurrencyWiseDetails).toEqual(currencySummaryData);
     expect(component.isForeignCurrency).toBe(false);
   });
 
-  it('ionViewWillEnter(): should update report details and currency and set Reimbursable to zero', () => {
+  it('ionViewWillEnter(): should update report details and currency and set Reimbursable amount', () => {
     component.view = ExpenseView.team;
     const erpt = {
       'Report Name': 'My Testing Report',
@@ -148,20 +148,20 @@ describe('FyViewReportInfoComponent', () => {
       'Created On': datePipe.transform(new Date('2022-10-31T13:54:46.317208'), 'MMM d, y'),
     };
     const paymentModeSummaryMock = {
-      ccc: {
-        name: 'example',
-        key: 'key123',
-        amount: 4600,
+      reimbursable: {
+        name: 'Reimbursable',
+        key: 'reimbursable',
+        amount: 207000.78,
         count: 200,
       },
     };
     spyOn(component, 'createEmployeeDetails');
     spyOn(component, 'getCCCAdvanceSummary');
     orgSettingsService.get.and.returnValue(of(orgSettingsRes));
-    transactionService.getCurrenyWiseSummary.and.returnValue(currencySummaryData);
-    transactionService.getPaymentModeWiseSummary.and.returnValue(paymentModeSummaryMock);
+    sharedExpensesService.getCurrenyWiseSummary.and.returnValue(currencySummaryData);
+    sharedExpensesService.getPaymentModeWiseSummary.and.returnValue(paymentModeSummaryMock);
     component.erpt$ = of(reportParam);
-    component.etxns$ = of(expenseList);
+    component.expenses$ = of(expenseResponseData2);
     fixture.detectChanges();
     component.ionViewWillEnter();
     expect(component.reportDetails).toEqual(erpt);
@@ -169,10 +169,10 @@ describe('FyViewReportInfoComponent', () => {
     expect(component.createEmployeeDetails).toHaveBeenCalledOnceWith(reportParam);
     expect(component.amountComponentWiseDetails).toEqual({
       'Total Amount': 46040,
-      Reimbursable: 0,
+      Reimbursable: 207000.78,
     });
     expect(component.getCCCAdvanceSummary).toHaveBeenCalledOnceWith(paymentModeSummaryMock, orgSettingsRes);
-    expect(transactionService.getCurrenyWiseSummary).toHaveBeenCalledOnceWith(expenseList);
+    expect(sharedExpensesService.getCurrenyWiseSummary).toHaveBeenCalledOnceWith(expenseResponseData2);
     expect(component.amountCurrencyWiseDetails).toEqual(currencySummaryData);
     expect(component.isForeignCurrency).toBe(false);
   });
